+++ cuda/exchange.cu
@@ -10,7 +10,7 @@ addexchange(float* __restrict__ Bx, float* __restrict__ By, float* __restrict__ 
             float* __restrict__ mx, float* __restrict__ my, float* __restrict__ mz,
             float* __restrict__ Ms_, float Ms_mul,
             float* __restrict__ aLUT2d, uint16_t* __restrict__ regions,
-            float wx, float wy, float wz, int Nx, int Ny, int Nz, uint8_t PBC) {
+            float wx, float wy, float wz, int Nx, int Ny, int Nz, uint16_t PBC) {
 
     int ix = blockIdx.x * blockDim.x + threadIdx.x;
     int iy = blockIdx.y * blockDim.y + threadIdx.y;
